+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+2.0.0
+===
+- Add parallel preloading capabilities.
+- Add asset resolver.
+- Add configuration options.
+
 1.0.0
 ===
 - Birthday!